agentic_architecture/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
rich>=13.0.0
pywinauto>=0.6.8
pywin32>=306
mcp>=0.1.0
click>=8.0.0 
google-genai==1.11.0
mcp==1.6.0
pydantic==2.11.3